erp/backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-cors-headers>=4.0
django-filter>=24.0
dj-database-url>=2.1
python-dotenv>=1.0
psycopg[binary]>=3.1
